pyproject.toml
[project]
name = "analyticaljunior"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "onnxruntime>=1.24.2",
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "rembg>=2.0.72",
]
